fix: Change UnitaryPrice type in t_articles table from float to decimal.

+++ Resto/files/Shop.sql
@@ -12,7 +12,7 @@ CREATE TABLE T_Articles (
 	IdArticle			int(4)		PRIMARY KEY AUTO_INCREMENT,
 	Description			varchar(30)	NOT NULL,
 	Brand				varchar(30)	NOT NULL,
-	UnitaryPrice		float(8)	NOT NULL DEFAULT 0
+	UnitaryPrice		DECIMAL(8, 2)	NOT NULL DEFAULT 0
 ) ENGINE = InnoDB;
 
 INSERT INTO T_Articles ( Description, Brand, UnitaryPrice ) VALUES ( 'Souris'     ,	'Logitoch', 65 );
